requirements.txt
aiohttp>=3.9.0
httpx>=0.25.0
pydantic>=2.5.0
rich>=13.7.0
beautifulsoup4>=4.12.0
python-dotenv>=1.0.0
fake-useragent>=1.4.0


